+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 Robofirm DevOps Tool: MySQL Database Support
 ============================================
 
+# 0.1.1
+- Removed DI config for \DevopsToolCore\Database\DatabaseMetadataProviderInterface
+- Added DatabaseAdapterFactory
+- Merged DatabaseMetadataProvider into DatabaseAdapter
+
 # 0.1.0
 - Initial build